Crate os_balloon

OpenStratos balloon software.

This crate provides the functionality required to control a stratospheric balloon. It provides several modules that can be enabled using cargo features, and it can be extended by adding more modules.

Example:

If you for example want to use the GPS and the GSM, but no real-time telemetry or Raspberry Pi camera, it's as simple as compiling the crate as follows:

cargo build --no-default-features --features="gps fona"

Here, the --no-default-features is required since by default, GPS, GSM (Adafruit FONA), Raspberry Pi camera and real-time transparent serial telemetry will be activated.

Configuration

OpenStratos is highly configurable. Please refer to the config module for further information.

Launcher

The project has a launcher in src/main.rs and can be launched by running cargo run. More information can be found in the launcher crate.

Simulation mode

In development…

Re-exports

pub use config::CONFIG;

pub use config::CONFIG;

Modules

config
Configuration module.

error
Error module.

fona
Adafruit FONA GSM module.

gps
GPS module.

logic
Logic module.

raspicam
Raspberry Pi camera module.

telemetry
Transparent serial telemetry module.

Constants

CONFIG_FILE
Configuration file.

STATE_FILE
Last state file, in the data directory.

Functions

generate_error_string
Generates a stack trace string of an error.

init_loggers
Initializes all loggers.

initialize_data_filesystem
Initializes the data file system for videos and images.

run
The main logic of the program.