Struct os_balloon::config::Battery

pub struct Battery { /* fields omitted */ }

Battery configuration structure.

Methods

impl Battery

pub fn main_min(self) -> f32

Gets the minimum voltage for the main battery when empty, at 0%, in volts (V).

pub fn main_max(self) -> f32

Gets the maximum voltage for the main battery when full, at 100%, in volts (V).

pub fn fona_min(self) -> f32

Gets the minimum voltage for the FONA battery when empty, at 0%, in volts (V).

pub fn fona_max(self) -> f32

Gets the maximum voltage for the FONA battery when full, at 0%, in volts (V).

pub fn main_min_percent(self) -> f32

Gets the minimum admisible percentage for main battery for the launch.

pub fn fona_min_percent(self) -> f32

Gets the minimum admisible percentage for FONA battery for the launch.