Struct os_balloon::config::Picture

pub struct Picture { /* fields omitted */ }

Picture configuration structure.

Methods

impl Picture

pub fn height(self) -> u16

Gets the configured picture height for the camera, in pixels.

pub fn width(self) -> u16

Gets the configured picture width for the camera, in pixels.

pub fn rotation(self) -> Option<u16>

Gets the configured picture rotation for the camera, in degrees (°).

pub fn quality(self) -> u8

Gets the configured picture quality for the camera.

pub fn exif(self) -> bool

Gets wether the camera should add available EXIF information to pictures.

pub fn raw(self) -> bool

Gets wether the camera should add raw sensor data to pictures as JPEG metadata.

pub fn exposure(self) -> Option<Exposure>

Gets the configured exposure for pictures.

pub fn brightness(self) -> Option<u8>

Gets the configured brightness for pictures.

pub fn contrast(self) -> Option<i8>

Gets the configured contrast for pictures.

pub fn sharpness(self) -> Option<i8>

Gets the configured sharpness for pictures.

pub fn saturation(self) -> Option<i8>

Gets the configured saturation for pictures.

pub fn iso(self) -> Option<u16>

Gets the configured ISO for pictures.

pub fn ev(self) -> Option<i8>

Gets the configured EV compensation for pictures.

pub fn white_balance(self) -> Option<WhiteBalance>

Gets the configured automatic white balance for pictures.

pub fn interval(self) -> u32

Gets the interval between pictures during flight.

pub fn repeat(self) -> Option<u32>

Gets the timeout for repeated picture.

Repeat each picture after these seconds (for issues with probe movement).

pub fn first_timeout(self) -> u32

Gets the timeout for first picture after launch, in seconds.