Struct os_balloon::config::Video

pub struct Video { /* fields omitted */ }

Video configuration structure.

Methods

impl Video

fn height(&self) -> u16

Gets the configured video height for the camera, in pixels.

fn width(&self) -> u16

Gets the configured video width for the camera, in pixels.

fn fps(&self) -> u8

Gets the configured video framerate for the camera, in frames per second.

fn bitrate(&self) -> u32

Gets the configured bitrate for videos.

fn exposure(&self) -> Option<Exposure>

Gets the configured exposure for videos.

fn brightness(&self) -> Option<u8>

Gets the configured brightness for videos.

fn contrast(&self) -> Option<i8>

Gets the configured contrast for videos.

fn sharpness(&self) -> Option<i8>

Gets the configured sharpness for videos.

fn saturation(&self) -> Option<i8>

Gets the configured saturation for videos.

fn iso(&self) -> Option<u16>

Gets the configured ISO for videos.

fn stabilization(&self) -> bool

Gets if video stabilization needs to be turned on.

fn ev(&self) -> Option<i8>

Gets the configured EV compensation for videos.

fn white_balance(&self) -> Option<WhiteBalance>

Gets the configured automatic white balance for videos.