Enum os_balloon::error::ErrorKind

pub enum ErrorKind {
    Msg(String),
    Io(Error),
    Toml(Error),
    Log(Error),
    LogSet(SetLoggerError),
    LogConfig(Errors),
    FromUTF8(FromUtf8Error),
    NulError(NulError),
    GPIO(Error),
    SerialPort(Error),
    GPSInvalidStatus(String),
    GPSInit,
    GpsAlreadyInitialized,
    ConfigOpen(PathBuf),
    ConfigRead(PathBuf),
    ConfigInvalidToml(PathBuf),
    ConfigInvalid(String),
    DataFSInit,
    LogAppender(&'static str),
    LogBuild,
    DirectoryCreation(PathBuf),
    LastStateRead,
    LastStateFileOpen,
    LastStateFileRead,
    LastStateFileWrite,
    InvalidState(String),
    Init,
    FonaInit,
    FonaNoEchoOff,
    FonaInitNoPowerOn,
    FonaNoSerial,
    FonaSerialEnd,
    FonaLongSms,
    FonaSmsAtCmgd,
    FonaSmsCmgs,
    FonaSmsOk,
    FonaSendCommandCrlf,
    FonaPartialResponse(String),
    FonaCommand,
    FonaCADCInvalidResponse,
    CameraAlreadyRecording,
    CameraFileExists(PathBuf),
    CameraTest,
    CameraTestRemove(PathBuf),
}

The kind of an error.

Variants
Msg(String)

A convenient variant for String.

Io(Error)

Standard library I/O error.

Toml(Error)

TOML deserializing error.

Log(Error)

Log error.

LogSet(SetLoggerError)

Error setting up logger.

LogConfig(Errors)

Logger configuration error.

FromUTF8(FromUtf8Error)

Error creating a String from UTF-8 data.

NulError(NulError)

A nul byte was found in the vector provided.

GPIO(Error)

A GPIO library error.

SerialPort(Error)

Serial port error

GPSInvalidStatus(String)

Invalid GPS status.

GPSInit

GPS initialization error.

GpsAlreadyInitialized

GPS was already initialized when trying to initialize it.

ConfigOpen(PathBuf)

Error opening configuration file.

ConfigRead(PathBuf)

Error reading configuration file.

ConfigInvalidToml(PathBuf)

Invalid TOML in configuration file.

ConfigInvalid(String)

Invalid configuration options.

DataFSInit

Error initializing the data filesystem.

LogAppender(&'static str)

Error creating a log appender.

LogBuild

Error building the logger.

DirectoryCreation(PathBuf)

Error creating a directory.

LastStateRead

Error reading last state.

LastStateFileOpen

Error opening last state file.

LastStateFileRead

Error reading last state file.

LastStateFileWrite

Error writing last state file.

InvalidState(String)

Invalid last state.

Init

Initialization error.

FonaInit

FONA initialization error.

FonaNoEchoOff

Error turning FONA echo off.

FonaInitNoPowerOn

FONA could not be turned on.

FonaNoSerial

No FONA when trying to send command.

FonaSerialEnd

FONA serial found EOF.

FonaLongSms

SMS was too long to be sent.

FonaSmsAtCmgd

Error sending SMS on 'AT+CMGD=1' response.

FonaSmsCmgs

Error reading +CMGS response.

FonaSmsOk

No OK received after sending SMS.

FonaSendCommandCrlf

Error reading CRLF (\r\n) after sendin command to FONA.

FonaPartialResponse(String)

FONA serial found EOF.

FonaCommand

Error sending command to FONA.

FonaCADCInvalidResponse

Invalid response to AT+CADC? (read ADC) command.

CameraAlreadyRecording

Camera was already recording.

CameraFileExists(PathBuf)

Camera output file already exists.

CameraTest

Camera testing error.

CameraTestRemove(PathBuf)

Error removing camera test file.

Methods

impl ErrorKind

fn description(&self) -> &str

A string describing the error kind.