Enum os_balloon::error::Fona

pub enum Fona {
    Init,
    PowerOn,
    EchoOff,
    NoSerial,
    SerialEnd,
    LongSms,
    SmsAtCmgf,
    SmsAtCmgs,
    SmsCmgs,
    SmsOk,
    LocAtCmgf,
    LocAtCgatt,
    LocAtSapbr,
    LocAtSapbrContype,
    LocAtSapbrApn,
    LocAtCipgsmloc,
    LocAtGprsDown,
    LocLon,
    LocLat,
    SendCommandCrlf,
    PartialResponse {
        response: String,
    },
    Command,
    CBCInvalidResponse,
    CADCInvalidResponse,
}

Errors related to the use of the Adafruit FONA module.

Variants
Init

Error initializing the FONA module.

PowerOn

Error turning the FONA module on.

EchoOff

Error turning the FONA module's "echo" functionality off.

NoSerial

There was no open serial connection when trying to send a command to the FONA module.

SerialEnd

EOF was found in the FONA serial.

LongSms

SMS was too long to be sent.

SmsAtCmgf

Error sending SMS on AT+CMGF=1 response.

SmsAtCmgs

Error sending AT+CMGS message when sending an SMS.

SmsCmgs

Error reading +CMGS response sending SMS.

SmsOk

No OK received after sending SMS.

LocAtCmgf

Error getting location on AT+CMGF=1 response.

LocAtCgatt

Error getting location on AT+CGATT=1 response.

LocAtSapbr

Error getting location on AT+SAPBR=1,1 response.

LocAtSapbrContype

Error getting location on AT+SAPBR=3,1,"CONTYPE","GPRS" response.

LocAtSapbrApn

Error getting location on AT+SAPBR=3,1,"APN","{fona.location_service}" response.

LocAtCipgsmloc

Error getting location on AT+CIPGSMLOC=1,1 response.

LocAtGprsDown

Error turning GPRS down.

LocLon

Error getting longitude via GPRS.

LocLat

Error getting latitude via GPRS.

SendCommandCrlf

Error reading CRLF (\r\n) after sending command to FONA.

PartialResponse

FONA serial found EOF.

Fields of PartialResponse

response: String

Contents of the partial response.

Command

Error sending command to FONA.

CBCInvalidResponse

Invalid response to AT+CBC (battery charge) command.

CADCInvalidResponse

Invalid response to AT+CADC? (read ADC) command.