Struct os_balloon::gps::Gps

pub struct Gps { /* fields omitted */ }

GPS information structure.

Methods

impl Gps

fn initialize(&mut self) -> Result<()>

Initializes the GPS.

fn fix_time(&self) -> DateTime<Utc>

Gets the time of the current fix.

fn status(&self) -> FixStatus

Gets the GPS fix status.

fn satellites(&self) -> u8

Gets the number of satellites connected.

fn latitude(&self) -> f32

Gets the latitude of the GPS antenna, in ° (degrees).

fn longitude(&self) -> f32

Gets the longitude of the GPS antenna, in ° (degrees).

fn altitude(&self) -> f32

Gets the altitude of the GPS antenna from sea level, in m.

fn pdop(&self) -> f32

Gets the position dilution of precision (3D).

fn hdop(&self) -> f32

Gets the horizontal dilution of precision (2D).

fn vdop(&self) -> f32

Gets the vertical dilution of precision (1D).

fn speed(&self) -> f32

Gets the speed of the velocity vector, in m/s.

fn course(&self) -> f32

Gets the course of the velocity vector, in ° (degrees).