Module os_balloon::logic

Logic module.

Structs

AcquiringFix
Acquiring fix state.

FixAcquired
Fix acquired state.

GoingDown
Going down state.

GoingUp
Going up state.

Init
Initialization state.

Landed
Landed state.

OpenStratos
Main OpenStratos state machine

SafeMode
Safe mode state.

ShutDown
Shut down state.

WaitingLaunch
Waiting launch state.

Enums

State
States of the onboard computer.

Traits

GetState
Trait to get the current state in the State enum for the current state in the state machine.

MainLogic
Trait implementing the main logic of the program.

StateMachine
Trait representing a state machine.

Functions

init
Initializes a new state machine.