Trait os_balloon::logic::GetState

pub trait GetState {
    fn get_state(&self) -> State;
}

Trait to get the current state in the State enum for the current state in the state machine.

Required Methods

fn get_state(&self) -> State

Gets the state enumeration variant for the current state.