Struct os_balloon::raspicam::Camera

pub struct Camera { /* fields omitted */ }

Camera structure.

This structure controlls the use of the camera.

Methods

impl Camera

fn record<T, P, FN>(&mut self, time: T, file_name: FN) -> Result<()> where
T: Into<Option<Duration>>,
P: AsRef<Path>,
FN: Into<Option<P>>,

Starts recording video with the camera.

If Some(Duration) is passed, the camera will stop recording after that time. This should be the default behaviour, and it will block for the given time. If no duration is given, it will record indefinitely, and the thread won't block. To stop the camera recording anytime, Camera::stop_recording() can be used.

An optional second parameter can be provided to specify a file name for the video recording, useful in case of testing. If that file name is provided, or if the method is executed as a test, the file will be removed after the recording, except if the mantain_test_video feature is used. If a file name is provided, a time should be provided too, and it will throw a warning if not.

Panics if the duration is less than 1 second.

fn stop_recording(&mut self) -> Result<()>

Stops the video recording.

fn is_recording(&self) -> bool

Checks if the camera is recording.

fn take_picture<P, FN>(&mut self, file_name: FN) -> Result<()> where
P: AsRef<Path>,
FN: Into<Option<P>>,

Takes a picture with the camera.