Struct os_balloon::config::Picture

pub struct Picture {
    height: u16,
    width: u16,
    rotation: Option<u16>,
    quality: u8,
    exif: Option<bool>,
    raw: Option<bool>,
    exposure: Option<Exposure>,
    brightness: Option<u8>,
    contrast: Option<i8>,
    sharpness: Option<i8>,
    saturation: Option<i8>,
    iso: Option<u16>,
    ev: Option<i8>,
    white_balance: Option<WhiteBalance>,
    interval: u32,
    repeat: Option<u32>,
    first_timeout: u32,
}

Picture configuration structure.

Fields
height: u16

Height of the picture, in px.

width: u16

Width of the picture, in px.

rotation: Option<u16>

Rotation of the camera, in degrees (°).

quality: u8

Quality of the picture, in px.

exif: Option<bool>

Wether to add EXIF data to pictures or not.

raw: Option<bool>

Wether to save the raw sensor data as JPG metadata.

exposure: Option<Exposure>

Exposure configuration.

brightness: Option<u8>

Brightness correction.

contrast: Option<i8>

Contrast correction.

sharpness: Option<i8>

Sharpness configuration.

saturation: Option<i8>

Saturation configuration.

iso: Option<u16>

ISO for the image.

ev: Option<i8>

EV correction configuration.

white_balance: Option<WhiteBalance>

White balance configuration.

interval: u32

Interval between pictures during flight.

repeat: Option<u32>

Repeat each picture after these seconds (for issues with probe movement).

first_timeout: u32

Timeout for first picture after launch, in seconds.

Methods

impl Picture

pub fn height(self) -> u16

Gets the configured picture height for the camera, in pixels.

pub fn width(self) -> u16

Gets the configured picture width for the camera, in pixels.

pub fn rotation(self) -> Option<u16>

Gets the configured picture rotation for the camera, in degrees (°).

pub fn quality(self) -> u8

Gets the configured picture quality for the camera.

pub fn exif(self) -> bool

Gets wether the camera should add available EXIF information to pictures.

pub fn raw(self) -> bool

Gets wether the camera should add raw sensor data to pictures as JPEG metadata.

pub fn exposure(self) -> Option<Exposure>

Gets the configured exposure for pictures.

pub fn brightness(self) -> Option<u8>

Gets the configured brightness for pictures.

pub fn contrast(self) -> Option<i8>

Gets the configured contrast for pictures.

pub fn sharpness(self) -> Option<i8>

Gets the configured sharpness for pictures.

pub fn saturation(self) -> Option<i8>

Gets the configured saturation for pictures.

pub fn iso(self) -> Option<u16>

Gets the configured ISO for pictures.

pub fn ev(self) -> Option<i8>

Gets the configured EV compensation for pictures.

pub fn white_balance(self) -> Option<WhiteBalance>

Gets the configured automatic white balance for pictures.

pub fn interval(self) -> u32

Gets the interval between pictures during flight.

pub fn repeat(self) -> Option<u32>

Gets the timeout for repeated picture.

Repeat each picture after these seconds (for issues with probe movement).

pub fn first_timeout(self) -> u32

Gets the timeout for first picture after launch, in seconds.