Struct os_balloon::config::Video

pub struct Video { /* fields omitted */ }

Video configuration structure.

Methods

impl Video

pub fn height(self) -> u16

Gets the configured video height for the camera, in pixels.

pub fn width(self) -> u16

Gets the configured video width for the camera, in pixels.

pub fn rotation(self) -> Option<u16>

Gets the configured picture rotation for the camera, in degrees (°).

pub fn fps(self) -> u8

Gets the configured video framerate for the camera, in frames per second.

pub fn bitrate(self) -> u32

Gets the configured bitrate for videos.

pub fn exposure(self) -> Option<Exposure>

Gets the configured exposure for videos.

pub fn brightness(self) -> Option<u8>

Gets the configured brightness for videos.

pub fn contrast(self) -> Option<i8>

Gets the configured contrast for videos.

pub fn sharpness(self) -> Option<i8>

Gets the configured sharpness for videos.

pub fn saturation(self) -> Option<i8>

Gets the configured saturation for videos.

pub fn iso(self) -> Option<u16>

Gets the configured ISO for videos.

pub fn stabilization(self) -> bool

Gets if video stabilization needs to be turned on.

pub fn ev(self) -> Option<i8>

Gets the configured EV compensation for videos.

pub fn white_balance(self) -> Option<WhiteBalance>

Gets the configured automatic white balance for videos.