Struct os_balloon::config::Video

pub struct Video {
    height: u16,
    width: u16,
    rotation: Option<u16>,
    fps: u8,
    bitrate: u32,
    exposure: Option<Exposure>,
    brightness: Option<u8>,
    contrast: Option<i8>,
    sharpness: Option<i8>,
    saturation: Option<i8>,
    iso: Option<u16>,
    stabilization: Option<bool>,
    ev: Option<i8>,
    white_balance: Option<WhiteBalance>,
}

Video configuration structure.

Fields
height: u16

Height of the video, in px.

width: u16

Width of the video, in px.

rotation: Option<u16>

Rotation of the camera, in degrees (°).

fps: u8

Frames per second (FPS) for the video.

bitrate: u32

Bit rate for the video, in bps (bits per second).

exposure: Option<Exposure>

Exposure configuration.

brightness: Option<u8>

Brightness correction.

contrast: Option<i8>

Contrast correction.

sharpness: Option<i8>

Sharpness configuration.

saturation: Option<i8>

Saturation configuration.

iso: Option<u16>

ISO for the image.

stabilization: Option<bool>

Video stabilization.

ev: Option<i8>

EV correction configuration.

white_balance: Option<WhiteBalance>

White balance configuration.

Methods

impl Video

pub fn height(self) -> u16

Gets the configured video height for the camera, in pixels.

pub fn width(self) -> u16

Gets the configured video width for the camera, in pixels.

pub fn rotation(self) -> Option<u16>

Gets the configured picture rotation for the camera, in degrees (°).

pub fn fps(self) -> u8

Gets the configured video framerate for the camera, in frames per second.

pub fn bitrate(self) -> u32

Gets the configured bitrate for videos.

pub fn exposure(self) -> Option<Exposure>

Gets the configured exposure for videos.

pub fn brightness(self) -> Option<u8>

Gets the configured brightness for videos.

pub fn contrast(self) -> Option<i8>

Gets the configured contrast for videos.

pub fn sharpness(self) -> Option<i8>

Gets the configured sharpness for videos.

pub fn saturation(self) -> Option<i8>

Gets the configured saturation for videos.

pub fn iso(self) -> Option<u16>

Gets the configured ISO for videos.

pub fn stabilization(self) -> bool

Gets if video stabilization needs to be turned on.

pub fn ev(self) -> Option<i8>

Gets the configured EV compensation for videos.

pub fn white_balance(self) -> Option<WhiteBalance>

Gets the configured automatic white balance for videos.