Struct os_balloon::gps::Gps

pub struct Gps {
    fix_time: DateTime<Utc>,
    status: FixStatus,
    satellites: u8,
    latitude: f32,
    longitude: f32,
    altitude: f32,
    pdop: f32,
    hdop: f32,
    vdop: f32,
    speed: f32,
    course: f32,
}

GPS information structure.

Fields
fix_time: DateTime<Utc>

Time of the current fix.

status: FixStatus

GPS fix status.

satellites: u8

Number of satellites connected.

latitude: f32

Latitude of the GPS antenna, in ° (degrees).

longitude: f32

Longitude of the GPS antenna, in ° (degrees).

altitude: f32

Altitude of the GPS antenna from sea level, in m.

pdop: f32

Position dilution of precision (3D).

hdop: f32

Horizontal dilution of precision (2D).

vdop: f32

Vertical dilution of precision (1D).

speed: f32

Speed of the velocity vector, in m/s.

course: f32

Course of the velocity vector, in ° (degrees).

Methods

impl Gps

pub fn initialize(&mut self) -> Result<(), Error>

Initializes the GPS.

pub fn fix_time(&self) -> DateTime<Utc>

Gets the time of the current fix.

pub fn status(&self) -> FixStatus

Gets the GPS fix status.

pub fn satellites(&self) -> u8

Gets the number of satellites connected.

pub fn latitude(&self) -> f32

Gets the latitude of the GPS antenna, in ° (degrees).

pub fn longitude(&self) -> f32

Gets the longitude of the GPS antenna, in ° (degrees).

pub fn altitude(&self) -> f32

Gets the altitude of the GPS antenna from sea level, in m.

pub fn pdop(&self) -> f32

Gets the position dilution of precision (3D).

pub fn hdop(&self) -> f32

Gets the horizontal dilution of precision (2D).

pub fn vdop(&self) -> f32

Gets the vertical dilution of precision (1D).

pub fn speed(&self) -> f32

Gets the speed of the velocity vector, in m/s.

pub fn course(&self) -> f32

Gets the course of the velocity vector, in ° (degrees).