Module os_balloon::logic

Logic module.

Modules

acquiring_fix
Acquiring fix logic.

fix_acquired
Fix acquired logic.

going_down
Going down logic.

going_up
Going up logic.

init
Initialization logic.

landed
Landed logic.

safe_mode
Safe mode logic.

shut_down
Shut down logic.

waiting_launch
Waiting launch logic.

Structs

AcquiringFix
Acquiring fix state.

CURRENT_STATE

FixAcquired
Fix acquired state.

GoingDown
Going down state.

GoingUp
Going up state.

Init
Initialization state.

Landed
Landed state.

OpenStratos
Main OpenStratos state machine

SafeMode
Safe mode state.

ShutDown
Shut down state.

WaitingLaunch
Waiting launch state.

Enums

State
States of the onboard computer.

Statics

CURRENT_STATE

Traits

GetState
Trait to get the current state in the State enum for the current state in the state machine.

MainLogic
Trait implementing the main logic of the program.

StateMachine
Trait representing a state machine.

Functions

init
Initializes a new state machine.

save_current_state
Saves the current state into the state file.