Enum os_balloon::logic::State

pub enum State {
    Init,
    AcquiringFix,
    FixAcquired,
    WaitingLaunch,
    GoingUp,
    GoingDown,
    Landed,
    ShutDown,
    SafeMode,
}

States of the onboard computer.

Variants
Init

Initialization.

AcquiringFix

Acquiring GPS fix.

FixAcquired

GPS fix has been acquired.

WaitingLaunch

Waiting for balloon launch.

GoingUp

Going up.

GoingDown

Going down.

Landed

Probe landed.

ShutDown

Shutting computer down.

SafeMode

Safe mode operation.

Methods

impl State

pub fn get_last() -> Result<Option<State>, Error>

Gets the last state of the application if there is one.

pub fn as_str(&self) -> &str

Gets the state as a string to be stored in the LAST_STATE file.